Refactoring AC code to a separate directory to simplify code structure.
