--- src/models.h
@@ -1,8 +1,0 @@
-#pragma once
-#include <string>
-#include <vector>
-
-struct KMer {
-    std::string value;
-    size_t length() const{ return value.size(); }
-};
