+++ src/parser.h
@@ -9,82 +9,9 @@
 #include "kseq.h"
 KSEQ_INIT(gzFile, gzread)
 
-#include "models.h"
 #include "kmers.h"
 #include "khash_utils.h"
 
-/// Record of one fasta sequence.
-struct FastaRecord {
-    // The name of the record. Starts with '>'.
-    std::string name;
-    // The genetic sequence.
-    std::string sequence;
-};
-
-/// Read fasta file with given path.
-std::vector<FastaRecord> ReadFasta(std::string &path) {
-    gzFile fp;
-    kseq_t *seq;
-    std::vector<FastaRecord> records;
-
-    fp = gzopen(path.c_str(), "r");
-    if (fp == nullptr) {
-        throw std::invalid_argument("couldn't open file " + path);
-    }
-    seq = kseq_init(fp);
-    while (kseq_read(seq) >= 0) {
-        records.push_back(FastaRecord{
-            seq->name.s,
-            seq->seq.s,
-        });
-    }
-    kseq_destroy(seq);
-    gzclose(fp);
-    return records;
-}
-
-/// Create a list of unique k-mers in no particular order.
-/// This runs in O(k*data.size) expected time.
-void AddKMersFromSequence(std::unordered_set<std::string> &kMers, std::string &data, int k) {
-    // Convert the sequence to uppercase letters.
-    std::transform(data.begin(), data.end(), data.begin(), toupper);
-    size_t possibleKMerEnd = k;
-    for (size_t i = 1; i <= data.size(); ++i) {
-        if (data[i-1] != 'A' && data[i-1] != 'C' && data[i-1] != 'G' && data[i-1] != 'T') {
-            // Skip this and the next k-1 k-mers.
-            possibleKMerEnd = i + k;
-        }
-        if (i >= possibleKMerEnd) {
-            kMers.insert(data.substr(i - k, k));
-        }
-    }
-}
-
-/// Return only the subset of k-mers where no two k-mers are complements of one another.
-/// The k-mers are chosen with no particular property.
-std::unordered_set<std::string> FilterKMersWithComplement(std::unordered_set<std::string> &kMers) {
-    std::unordered_set<std::string> ret;
-    for (auto &&x : kMers) {
-        auto kMer = KMer{x};
-        if (ret.count(ReverseComplement(kMer).value) == 0) ret.insert(x);
-    }
-    return ret;
-}
-
-/// Create a list of unique k-mers in no particular order.
-/// This runs in O(k*data.size) expected time.
-std::vector<KMer> ConstructKMers(std::vector<FastaRecord> &data, int k, bool complements) {
-    std::unordered_set<std::string> uniqueKMers;
-    for (auto &&record : data) {
-        AddKMersFromSequence(uniqueKMers, record.sequence, k);
-    }
-    if (complements) uniqueKMers = FilterKMersWithComplement(uniqueKMers);
-    std::vector<KMer> result;
-    for (auto &kMer : uniqueKMers) {
-        result.push_back(KMer{kMer});
-    }
-    return result;
-}
 
 /// Fill the k-mer dictionary with k-mers from the given sequence.
 /// If complements is true, always add the canonical k-mers.
@@ -160,4 +87,10 @@ void AssertEOF(kseq_t *seq, std::string message) {
     if (kseq_read(seq) >= 0) {
         throw std::invalid_argument(message);
     }
+}
+
+/// Print the fasta file header.
+void WriteName(const int k, std::ostream &of) {
+    of << ">superstring ";
+    of << "k=" << k << std::endl;
 }
