+++ tests/streaming_unittest.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "../src/streaming.h"
+#include "../src/ac/streaming.h"
 #include "gtest/gtest.h"
 namespace {
